+++ sydtest-mutation-plugin/src/Test/Syd/Mutation/Plugin/Operator/Arith.hs
@@ -4,7 +4,7 @@ module Test.Syd.Mutation.Plugin.Operator.Arith (theOperator) where
 
 import Control.Monad.Reader (ask)
 import qualified Data.Text as T
-import Test.Syd.Mutation.Plugin.Instrument (InstrM, InstrumentEnv (..), MutationAlt (..), MutationOperator (..), SrcSpanDelta (..), liftTcM)
+import Test.Syd.Mutation.Plugin.Instrument (InstrM, InstrumentEnv (..), MutationAlt (..), MutationOperator (..), MutationOperatorKind (..), SrcSpanDelta (..), liftTcM)
 import Test.Syd.Mutation.Plugin.Operator.Util (TcOpApp (..), matchTcOpApp, mkOpReplacement)
 
 theOperator :: MutationOperator
@@ -12,7 +12,7 @@ theOperator =
   MutationOperator
     { operatorName = "Arith",
       operatorDescription = "Replace any binary arithmetic operator with every other arithmetic operator",
-      operatorMatch = \le -> case matchTcOpApp le of
+      operatorKind = ExpressionOperator $ \le -> case matchTcOpApp le of
         Just tcOp@(TcOpApp {tcOpAppOcc = occ})
           | occ `elem` arithOps ->
               Just (action tcOp)