+++ sydtest-mutation-plugin/src/Test/Syd/Mutation/Plugin/Operator/IntLit.hs
@@ -5,7 +5,7 @@ module Test.Syd.Mutation.Plugin.Operator.IntLit (theOperator) where
 import qualified Data.Text as T
 import GHC
 import GHC.Types.SourceText (il_value)
-import Test.Syd.Mutation.Plugin.Instrument (InstrM, MutationAlt (..), MutationOperator (..), SrcSpanDelta (..))
+import Test.Syd.Mutation.Plugin.Instrument (InstrM, MutationAlt (..), MutationOperator (..), MutationOperatorKind (..), SrcSpanDelta (..))
 import Test.Syd.Mutation.Plugin.Operator.Util (mkIntLitExpr)
 
 theOperator :: MutationOperator
@@ -13,7 +13,7 @@ theOperator =
   MutationOperator
     { operatorName = "IntLit",
       operatorDescription = "Replace an integer literal n with 0, 1, and -n",
-      operatorMatch = \case
+      operatorKind = ExpressionOperator $ \case
         (L _ (HsOverLit _ (OverLit oltc@(OverLitTc {ol_type = ty}) (HsIntegral il)))) ->
           Just (action ty oltc (il_value il))
         _ -> Nothing